Compact json serialization for efficiency

+++ src/main/scala/org/huwtl/penfold/readstore/NewEventsProvider.scala
@@ -6,7 +6,7 @@ class NewEventsProvider(nextExpectedEventProvider: NextExpectedEventIdProvider, 
     val lastEventId = eventStoreQueryService.retrieveIdOfLast getOrElse EventSequenceId(-1)
 
     for {
-      eventId <- (nextExpectedEventId.value.toInt to lastEventId.value.toInt).toStream
+      eventId <- (nextExpectedEventId.value to lastEventId.value).toStream
       event <- eventStoreQueryService.retrieveBy(EventSequenceId(eventId))
     } yield event
   }
